requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.5.0
sqlalchemy>=2.0.35
alembic>=1.13.1
psycopg2-binary>=2.9.10
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1,<4.1.0
python-multipart>=0.0.9
pgvector>=0.3.0
openai>=1.12.0
voyageai>=0.2.1
redis>=5.0.1
celery>=5.3.6
boto3>=1.34.0
pinecone>=5.0.0
python-docx>=1.1.0
pypdf>=4.0.1
pdfplumber>=0.11.4
openpyxl>=3.1.5
tiktoken>=0.9.0
chardet>=5.2.0
rank-bm25>=0.2.2
pytesseract>=0.3.13
pdf2image>=1.17.0
Pillow>=11.1.0
striprtf>=0.0.26
beautifulsoup4>=4.12.3
lxml>=5.3.0
tenacity>=8.2.3
prometheus-client>=0.21.0
dnspython>=2.6.1
email-validator
aiofiles>=23.2.1
python-magic>=0.4.27
reportlab>=4.1.0
sendgrid>=6.11.0
pycryptodome>=3.20.0

# ── Phase 3: 進階文檔解析 ──
jieba>=0.42.1
llama-cloud-services>=0.6.3
llama-parse>=0.6.90  # backward compat wrapper, pulls in llama-cloud-services
llama-index-core>=0.14.0
python-pptx>=1.0.0
trafilatura>=2.0.0
nest-asyncio>=1.6.0

# ── Phase 12: LLMOps 可觀測性 ──
langfuse>=2.0.0

# ── Phase 12: MCP Server ──
mcp>=1.0.0
